app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask_migrate==4.1.0
flask_marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.10